type bug

+++ macow/data/image.py
@@ -156,11 +156,11 @@ def preprocess(img, n_bits, beta, nsamples=1):
     if beta is not None:
         # add noise (-0.5, 0.5)
         if nsamples == 1:
-            u = beta.rsample(img.size()) - 0.5
+            u = beta.rsample(img.size()).type_as(img) - 0.5
             img = img + u
         else:
             batch, c, h, w = img.size()
-            u = beta.rsample((batch, nsamples, c, h, w)) - 0.5
+            u = beta.rsample((batch, nsamples, c, h, w)).type_as(img) - 0.5
             img = img.unsqueeze(1) + u
     # normalize
     img = img.div(n_bins)
